# [Enhancement]lazy decode for parquet reader lazy columns