@eshishki 2024-05-31
i tested this commit with my sample production queries
```
with
curr_transactions as (
    select * from transactions
),
curr_acc_hist as (
    select * from accounts_history
),
curr_accounts as (
    select * from accounts
),
acc_hist_shortcut as (
    SELECT
        aa.tr_id,
        aa.operation_type,
        a.shortcut,
        aa.account_id
    FROM curr_acc_hist aa
    INNER JOIN curr_accounts a on aa.account_id = a.id

),
transactions as (
    select *,
        case when dest_acc = ah.account_id or dest_acc_name = ah.shortcut then 'income'
        else 'expense' end as transaction_type
    from curr_transactions tr join acc_hist_shortcut ah ON tr.id = ah.tr_id
    where status = 'approved'
    AND convert_tz(created_at,'Etc/UTC', 'Asia/Dushanbe')
    between '2023-02-01' and '2023-02-11'

)
select * from transactions limit 100;
``` 
gives 
```
ERROR 1064 (HY000): FileReader::get_next failed. reason = Internal error: DictDecoder append strings to column failed
be/src/formats/parquet/stored_column_reader.cpp:338 _read_values_on_levels(num_values, content_type, dst, false)
be/src/formats/parquet/complex_column_reader.cpp:198 _child_readers[field_name]->read_range(range, filter, child_column)
be/src/formats/parquet/group_reader.cpp:184 _column_readers[slot_id]->read_range(range, filter, (*chunk)->get_column_by_slot_id(slot_id))
be/src/formats/parquet/group_reader.cpp:1
```

i couldn't produce a self contained test case unfortunately
but this query works in 3.2.7 and fails on main branch with lazy decode