@zombee0 2024-06-07
@eshishki  could you check the be/log if there is more log message? and could you check this query just without this pr?