@zombee0 2024-06-18
> @zombee0 i have verified that the issue persist in main branch and absent in branch 3.3 (without this pr), i've also verified that there is no issue up to this commit

will be fixed by https://github.com/StarRocks/starrocks/pull/47147, thank you!